+++ src/Rocketeer/Services/Ignition/Plugins.php
@@ -57,8 +57,8 @@ class Plugins
 	public function findPackageConfiguration($package)
 	{
 		$paths = array(
-			$this->paths->getBasePath().'vendor/%s/src/config',
-			$this->paths->getBasePath().'vendor/%s/config',
+			$this->paths->getApplicationPath().'vendor/%s/src/config',
+			$this->paths->getApplicationPath().'vendor/%s/config',
 			$this->paths->getUserHomeFolder().'/.rocketeer/vendor/%s/src/config',
 			$this->paths->getUserHomeFolder().'/.rocketeer/vendor/%s/config',
 			$this->paths->getUserHomeFolder().'/.composer/vendor/%s/src/config',