+++ tests/Abstracts/Strategies/AbstractStrategyTest.php
@@ -8,6 +8,9 @@ class AbstractStrategyTest extends RocketeerTestCase
 	public function testCanCheckForManifestWithoutServer()
 	{
 		$this->app['path.base'] = realpath(__DIR__.'/../../..');
+		$this->swapConfig(array(
+			'rocketeer::paths.app' => realpath(__DIR__.'/../../..'),
+		));
 
 		$this->usesComposer(false);
 		$strategy = $this->builder->buildStrategy('Dependencies', 'Composer');